requirements.txt
psycopg2
requests
minio
python-dotenv